Adding full state observable (#15)
* full state observation operator

* some minor updates to other files, __init__, and adding license boiler plate

* removed adjmult method from full space operator and ported functionality to transpmult

* tabs and spaces issue

* combined  with  in full state observation operator

* tab vs spaces issue

* compactifying

* no more state space observable

* adding a unit test and modifying some imports

* updating some comments
